% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_all.R
\name{calib_em_all}
\alias{calib_em_all}
\title{Calibrate the model for all watersheds}
\usage{
calib_em_all(data_obs)
}
\arguments{
\item{data_obs}{List with input and evaluation data for a set of watersheds}
}
\description{
Calibrate the model for all watersheds
}

