% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{calib_wrapper_gr4j}
\alias{calib_wrapper_gr4j}
\title{GR4J calibration wrapper}
\usage{
calib_wrapper_gr4j(Param, indata, evaldata)
}
\arguments{
\item{Some}{inputs}
}
\value{
Some outputs
}
\description{
Wrapper for running GR4J and computing performance measures
}

