% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{calib_wrapper_model}
\alias{calib_wrapper_model}
\title{Model (gr4j+snow) calibration wrapper}
\usage{
calib_wrapper_model(Param, indata, evaldata)
}
\arguments{
\item{Some}{inputs}
}
\value{
Some outputs
}
\description{
Wrapper for running GR4J and snow model and computing performance measures
}

