% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{gr4j_test_run}
\alias{gr4j_test_run}
\title{Test GR4J}
\usage{
gr4j_test_run()
}
\arguments{
\item{Some}{inputs}
}
\value{
Some outputs
}
\description{
Run GR4J and check against results from airGR package
}

