% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{model_test_calib}
\alias{model_test_calib}
\title{Test calibration of snow and hydrological model}
\usage{
model_test_calib()
}
\description{
Calibrate the models and check results against observations
}

