% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrapper.R
\name{model_wrapper}
\alias{model_wrapper}
\title{Wrapper for running models}
\usage{
model_wrapper(indata)
}
\arguments{
\item{Some}{inputs}
}
\value{
Some outputs
}
\description{
Wrapper for running models
}

