% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrapper_ens.R
\name{model_wrapper_ens}
\alias{model_wrapper_ens}
\title{Wrapper for running the model for filter algorithms}
\usage{
model_wrapper_ens(indata)
}
\description{
Wrapper for running the model for filter algorithms
}

