% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Interactive plot of model results}
\usage{
plot_interactive(data_obs, data_sim)
}
\arguments{
\item{data_obs}{Observed data (in format given by package NVEDATA)}

\item{data_mod}{Model data (in raw output format)}
}
\description{
Interactive plot of model results
}
\examples{

iwsh <- 3

data_obs <- sample_data[[iwsh]]

indata = list(Time           = data_obs$time_vec,
              Prec           = data_obs$Prec,
              Tair           = data_obs$Tair,
              PET            = rep(0, nrow(data_obs$Prec)),
              SWE            = matrix(0, nrow = 1, ncol = ncol(data_obs$Prec)),
              St             = matrix(0, nrow = 2, ncol = 1),
              StUH1          = matrix(0, 20, ncol = 1),
              StUH2          = matrix(0, 40, ncol = 1),
              Param          = c(74.59, 0.81, 214.98, 1.24, 3.69, 1.02),
              frac_elev_band = data_obs$frac_elev_band)

data_sim <- model_wrapper(indata)

plot_interactive(data_obs, data_sim)
}

