% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotting_results}
\alias{plotting_results}
\title{Plotting model results}
\usage{
plotting_results(data_obs, data_mod, path)
}
\arguments{
\item{data_obs}{Observed data (in format given by package NVEDATA)}

\item{data_mod}{Model data (in raw output format)}

\item{path}{Path for saving plots}
}
\description{
Plotting model results
}

