% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_simulations.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensemble for one watershed}
\usage{
run_ensemble(data_obs, param, NEns)
}
\description{
Run ensemble for one watershed
}
\examples{

NEns <- 10

iwsh <- 3

data_obs <- sample_data[[iwsh]]

param <- c(74.59, 0.81, 214.98, 1.24, 3.69, 1.02)

Qsim <- run_ensemble(data_obs, param, NEns)

plot(rowMeans(Qsim),type = "l")
lines(apply(Qsim,1,min))
lines(apply(Qsim,1,max))

}

