% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrapper.R
\name{snow_wrapper}
\alias{snow_wrapper}
\title{Wrapper for running snow model}
\usage{
snow_wrapper(insnow)
}
\arguments{
\item{indata}{A list with the following items:

\itemize{
  \item \code{Prec} Matrix with precipitation, dimensions Ntimes * NZones
  \item \code{Tair} Matrix with air temperature, dimensions Ntimes * NZones
  \item \code{SWE} Initial state of snow water equivalent
  \item \code{Param} Model parameters
}}
}
\value{
Simulated snow water equivalents
}
\description{
The snow model is a simple temperature index model.
}
\examples{

iwsh <- 3

indata = list(Prec   = sample_data[[iwsh]]$Prec,
              Tair   = sample_data[[iwsh]]$Tair,
              SWE    = matrix(0, nrow = 1, ncol = ncol(sample_data[[iwsh]]$Prec)),
              Param  = 3)

res_sim <- snow_wrapper(indata)

}

