% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calib_wrapper_gr4j}
\alias{calib_wrapper_gr4j}
\title{Calibration wrapper for hydrological model (GR4J)}
\usage{
calib_wrapper_gr4j(Param, indata, evaldata)
}
\arguments{
\item{Param}{Vector with parameter values}

\item{indata}{List with required inputs (excluding Param)}

\item{evaldata}{Runoff data for validation}
}
\value{
Goodness of fit measure
}
\description{
Wrapper for running GR4J and computing performance measures
}

