% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calib_wrapper_model}
\alias{calib_wrapper_model}
\title{Calibration wrapper for snow and hydrological model}
\usage{
calib_wrapper_model(Param, indata, evaldata)
}
\arguments{
\item{Param}{Vector with parameter values}

\item{indata}{List with required inputs (excluding Param)}

\item{evaldata}{Runoff data for validation}
}
\value{
Goodness of fit measure
}
\description{
Wrapper for running snow module and GR4J and computing performance measures
}

