% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{model_test_calib}
\alias{model_test_calib}
\title{Test GR4J and snow model}
\usage{
model_test_calib()
}
\arguments{
\item{Some}{inputs}
}
\value{
Some outputs
}
\description{
Calibrate GR4J with snow model
}

